Reimplementation of caching with lifespan and threads. Web.py route removed.
